Fix image caching and docker

+++ audioboos-server/Services/Images/ImageCacher.cs
@@ -19,10 +19,21 @@ public static class ImageCacher {
             Directory.CreateDirectory(Path.GetDirectoryName(cacheFile));
         }
 
-        var bufferFile = await HttpHelpers.DownloadFile(imageFile, Path.GetTempFileName());
-        using var image = await Image.LoadAsync(bufferFile);
-        await image.SaveAsJpegAsync(cacheFile);
-        File.Delete(bufferFile);
+        if (imageFile.StartsWith("http")) {
+            //image is a remote url
+            var bufferFile = await HttpHelpers.DownloadFile(imageFile, Path.GetTempFileName());
+            using var image = await Image.LoadAsync(bufferFile);
+            await image.SaveAsJpegAsync(cacheFile);
+            File.Delete(bufferFile);
+        } else {
+            //image is a local file in the audio folder
+            File.Copy(
+                //TODO: hack for using live db in debug
+                Constants.DebugMode
+                    ? imageFile.Replace("/audio/", "/mnt/frasier/audio/MuziQ/Pedestrian/Store/")
+                    : imageFile,
+                cacheFile);
+        }
 
         return File.Exists(cacheFile);
     }
